backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
google-cloud-storage>=2.14.0
httpx>=0.27.0
python-multipart>=0.0.12
pydantic>=2.0.0
pydantic-settings>=2.0.0
